Add d_type to the KBDIR entry structure.
On systems that support it, and in evnrionments, where it makes sense,
d_type will be filled with _FILE, _DIR, _UNKNOWN or _UNIXY (for sockets,
pipes, etc).

For all other situations, _UNKNOWN must be set, and additional tests
are to be performed.

Currently, only Linux/BSD with native filesystems are supported.